requirements.txt
pyspark==3.5.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
numpy>=1.20.0
kafka-python>=2.0.2
vaderSentiment>=3.3.2
zstandard>=0.19.0
pyarrow>=10.0.0
scikit-learn>=1.0.0
tqdm>=4.62.0
